requirements.txt
# --- Core Application ---
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0

# --- Database ---
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# --- AI & APIs ---
deepgram-sdk==3.1.3
google-generativeai==0.4.0
PyYAML==6.0.1

# --- Google Calendar API ---
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
